+++ LookupDropdown/ControlManifest.Input.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <manifest>
-  <control namespace="DR" constructor="LookupDropdown" version="0.0.91" display-name-key="LookupDropdown" description-key="PCF to render a lookup field as a dropdown. (Optional) Include record image, Customize record text, Open record button." control-type="standard" preview-image="img/lookupdropdown.png">
+  <control namespace="DR" constructor="LookupDropdown" version="0.0.94" display-name-key="LookupDropdown" description-key="PCF to render a lookup field as a dropdown. (Optional) Include record image, Customize record text, Open record button." control-type="standard" preview-image="img/lookupdropdown.png">
 
     <property name="lookupfield" display-name-key="Lookup Field" description-key="Lookup field to render as dropdown" of-type="Lookup.Simple" usage="bound" required="true" />
     <property name="customtext" display-name-key="Custom Text" description-key="(Optional) customize display text with any text attributes of the record. put attribute schema names between curly braces {}. Ex. {new_firstname} {new_lastname}. Leave blank to show record primaryname. See documentation for multi-language support." of-type="SingleLine.Text" usage="input" />
