+++ LookupDropdown/hooks/useRecords.ts
@@ -1,4 +1,3 @@
-import { IDropdownOption } from '@fluentui/react/lib/Dropdown'
 import { useQuery } from '@tanstack/react-query'
 
 import { usePcfContext } from '../services/PcfContext'
@@ -13,7 +12,7 @@ export const useRecords = () => {
   const { data, status, error, isFetching } =
     useQuery<ComponentFramework.WebApi.Entity[], Error>(
       {
-        queryKey: ['lookuprecords', pcfcontext.instanceid, pcfcontext.dependentValue?.id],
+        queryKey: ['lookuprecords', pcfcontext.instanceid, pcfcontext.dependentValue?.id, pcfcontext.viewid],
         queryFn: () => pcfcontext.getLookupRecords(entityname, primaryid, primaryname, primaryimage, fetchxml, metadata!),
         enabled: !!entityname && !!primaryid && !!fetchxml,
         staleTime: Infinity
@@ -23,31 +22,6 @@ export const useRecords = () => {
   return { records: data, status,
     error,
     isFetching }
-}
-
-export const useRecordsAsOptions = () => {
-  const pcfcontext = usePcfContext()
-  const { records, status, error, isFetching } = useRecords()
-  const { entityname } = useLookupView()
-  const { primaryid, primaryname, primaryimage } = useMetadata(entityname)
-
-  const options:IDropdownOption[] = records
-    ? [{ key: -1, text: pcfcontext.SelectText() }].concat(records.map(e => {
-        const imagesrc = e?.[primaryimage] == null
-          ? undefined
-          : `data:image/jpeg;base64,${e?.[primaryimage]}`
-        return {
-          key: e[`${primaryid}`],
-          text: pcfcontext.getRecordText(e, primaryname),
-          data: {
-            imagesrc: imagesrc,
-            recordname: e[`${primaryname}`]
-          }
-        }
-      }))
-    : [{ key: -1, text: pcfcontext.SelectText() }]
-
-  return { options, status, error, isFetching }
 }
 
 
@@ -65,13 +39,13 @@ export const useTagPickerOptions = () => {
   const { primaryid, primaryname, primaryimage } = useMetadata(entityname)
 
   const options:IRecord[] = records ? [{ id: '-1', displaytext: pcfcontext.SelectText() }].concat(records?.map(e => {
-        const imagesrc = e?.[primaryimage] == null
+        const imagesrc = e?.[primaryimage] == null || pcfcontext.isMasked
           ? undefined
           : `data:image/jpeg;base64,${e?.[primaryimage]}`
         return {
           id: e[`${primaryid}`],
           primaryname: e[`${primaryname}`],
-          displaytext: pcfcontext.getRecordText(e, primaryname),
+          displaytext: pcfcontext.isMasked ? '********' : pcfcontext.getRecordText(e, primaryname),
           imagesrc: imagesrc
         }
       })) : [{ id: '-1', displaytext: pcfcontext.SelectText() }]
