+++ package.json
@@ -14,9 +14,8 @@
     "deploy-storybook": "gh-pages -d storybook-static"
   },
   "dependencies": {
-    "@fluentui/react": "^8.117.0",
-    "@fluentui/react-components": "^9.54.0",
-    "@tanstack/react-query": "^5.40.1",
+    "@fluentui/react-components": "^9.54.2",
+    "@tanstack/react-query": "^5.45.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "uuid": "^9.0.1"
